requirements.txt
discord
d20
python-dotenv
pandas
fastapi
pydantic
gspread